Feat/users correction

+++ src/services/UserService.ts
@@ -1,8 +1,8 @@
 import { validate } from "class-validator";
 import { AppDataSource } from "../data-source";
-import { User } from "../entities/User";
+import { User, UserRole } from "../entities/User";
 import { formatErrors } from "../helpers/formatErrors";
-import { ApiError, BadRequestError, NotFoundError } from "../helpers/apiError";
+import { ApiError, BadRequestError, NotFoundError, UnauthorizedError } from "../helpers/apiError";
 import bcrypt from "bcryptjs";
 
 
@@ -26,11 +26,14 @@ export class UserService{
         if (existingUser) {
             throw new BadRequestError("Email fornecido já está em uso!");
         }
+        if(userData.role === UserRole.ADMIN){
+            throw new UnauthorizedError("Acesso negado. Apenas administradores podem criar outros administradores.");
+        }
         
         const hashedPassword = await bcrypt.hash(userData.password!, 10);
     return await this.userRepository.save({
       ...userData,
-      password: hashedPassword,
+      password: hashedPassword
     });
     };
 
@@ -39,20 +42,58 @@ export class UserService{
     };
 
     delete = async (userId: number)=>{
-        const user = this.userRepository.findOne({ where:{idUser: userId}});
+        const user = await this.userRepository.findOne({ where:{idUser: userId}});
+        console.log(user)
+        if (!user) {
+            throw new NotFoundError("Usuario não encontrado.")
+        }
+        if (user.isActive === true) {
+            throw new BadRequestError("Usuario ativo não pode ser excluído.")
+        }
+        if (user.task && user.task.length > 0) {
+            throw new BadRequestError("Usuario com tarefas associadas não pode ser excluído.")
+        }
         return await this.userRepository.delete(userId);
     };
 
-    update = async(userId: number, data: Partial<User>)=>{
+    update = async(userId: number, data: Partial<User>, requestingUserId: number, userRole: UserRole)=>{
         const user = await this.userRepository.findOne({where:{idUser: userId}});
         if(!user){
             throw new NotFoundError("Usuario não encontrado.")
         }
+        
+        if((data.name || data.lastName || data.email || data.password) && userId !== requestingUserId){
+            throw new UnauthorizedError("Acesso negado. Você não pode modificar informações pessoais de outros usuários.");
+        }
+        if(data.email){
+            const existingUser = await this.userRepository.findOneBy({
+            email: data.email,
+            });
+            if(existingUser && existingUser.idUser !== userId) {
+                throw new BadRequestError("Email fornecido já está em uso!");
+            }
+        }
+        if(data.password){
+            const hashedPassword = await bcrypt.hash(data.password, 10);
+            data.password = hashedPassword;
+        }
+        if(data.role){
+            if(user.role !== UserRole.ADMIN){
+                throw new UnauthorizedError("Acesso negado. Apenas administradores podem modificar o cargo de um usuário.");
+            }
+            if(user.role === UserRole.ADMIN && requestingUserId !== userId){
+                throw new UnauthorizedError("Acesso negado. Administradores não podem modificar o cargo de outros administradores.");
+            }
+        }
+        await this.validateSchema(data, true);
         this.userRepository.merge(user, data)
         return await this.userRepository.save(user)
     };
 
-    toggleActive= async(userId: number)=>{
+    toggleActive = async(userId: number, userRole: UserRole)=>{
+        if (userRole !== UserRole.ADMIN) {
+            throw new UnauthorizedError("Acesso negado. Apenas administradores podem ativar ou desativar usuários.");
+          }
         const user = await this.userRepository.findOne({where:{idUser: userId}});
         if(!user){
             throw new NotFoundError("Usuario não encontrado.")
